Release 3.0.0 (#87)
* created email layout builder (back) and confirmation page (front)

* Changed from multiple to one function for navigation in home page

* Added condition to show edit button in branches page

* fixed path to branches

* Changed menu-icons visibility

* Set border and outline to none when readonly is on

* Created page to redefine password

* Added custom variables to card-create-sku.component.css

* Added card-create-sku custom values in measu-unit page

* Creating two 404 pages (#51)

* Fixed readonly logic in modal-show-more and added feature to modal-update-measu-unit

* Added condition to show add-branch button in the branches page

* Created method for inserting and removing products

* Changed button text in card-measu-unit

* Fixed user-card path

* Feature/notifications (#54)

* adds notification to category components

* refactors confirmation dialog

* adds notification to login

* Fixed readonly logic in measu-unit and branch components and added edit button to add modal-update-sku

* Added edit button feature to category-update-form component

* Fixed application-dev misstype

* Resolved conflicts with Email-confirmation merge

* Returning pageable for user filtered request

* Added edit button feature to modal-update-client component

* adds pagination to category and warehouse (#55)

* added readonly condition to modal-update-client

* Created transaction history

* Created transaction history (#57)

* adds controller to findAll without pagination (#58)

* creating create transaction page (#59)

* Generated warehouse-slot page and card-warehouse components

* warehouse slot page layout - first version

* Feature/category and warehouseSlot pagination (#60)

* adds MatPaginator

* adds pagination to category and warehouseSlot

* added link to warehouse-slots from the home page and header component

* Feature/pagination clientsand SKU (#61)

* finished

* Added endpoint in transaction for front end use

* Feature/enhance jwt (#62)

* adding field to jwt

* removing additional information from respose body

* Transaction history cards (#63)

* Created transaction history

* Creating cards to show transaction-history

* fixed responsiveness in measu-unit page and warehouse-slot page and card component

* fixed responsiveness in warehouse-slot page

* What was broken, now is mended

* minor tweaks in client autocomplete

* layout changes to card-warehouse-slot

* Implemented create new user modal

* Removed duplicated line at app.module.ts

* Fixing client inventory and transaction history (#66)

* Created transaction history

* Fixing Client-inventory and Transactions-history

* its the little things of life

* Added state select and zip code mask to modal-add-branch

* Added state select and zip code mask to modal-show-more branch

* Added date mask pipe to card-warehouse-slot

* updating auth service (#67)

* Finished endpoint for ws

* Fixed zip code mask

* adds notification to most services (#69)

* fixes duplication bug

* fixes spelling mistake

* Client now dont use window reload

* Implemented forgot password methods

* Changing from 'Warehouse Slots' to 'Warehouse'

* Style fixing (#71)

* Created transaction history

* Fixing animation

* Hidinig arrows

* Feature/implementing create transaction (#72)

* connecting front with back end

* updating toasts

* search of warehouseslot (#73)

* adding transactions permissions

* changing application

* transaction create now sends DTo

* adds notification to client (#74)

* Creating search in warehouseSlots (#75)

* Just minor tweaks on the ws response

* adds spinner to client creation

* updating sql image

* Feature/transactions exceptions (#76)

* hadling exception

* changing toast message

* Just minor tweaks on the sku error trowing

* Fixing card to take 5 digits

* Just minor tweaks on the sku error trowing

* Fixing aisle and bay inputs, and accepting 5 digits

* Fixing search in warehouse-slot (#78)

* Fixed create branch logic

* updating styles for login and confirmation pages (#79)

* Fixed div

* handles deleting violations (#81)

* Presentation of 08-12 (#82)

* teste

* te

* dont know why

* Update README.md (#84)

* Update README.md

* Fixing minors issues

* Update README.md

* Update README.md

* Fixing card-warehouse-slot layout

* Fixed some userControllerTest methods

* Fixing toastr message

* minor fixer

* Update README.md

* Changes made on application.yml

* Fixing minors issues

* Update README.md

+++ wms/src/main/java/com/twms/wms/services/MeasurementUnitService.java
@@ -1,7 +1,9 @@
 package com.twms.wms.services;
 
 import com.twms.wms.entities.MeasurementUnit;
+import com.twms.wms.entities.SKU;
 import com.twms.wms.repositories.MeasurementUnitRepository;
+import com.twms.wms.repositories.SKURepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -9,6 +11,7 @@ import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 import java.util.Optional;
 
@@ -17,6 +20,10 @@ public class MeasurementUnitService {
 
     @Autowired
     MeasurementUnitRepository measurementUnitRepository;
+
+    @Autowired
+    SKURepository skuRepository;
+
     public Page<MeasurementUnit> read(Pageable pageable) {
         return  measurementUnitRepository.findAll(pageable);
     }
@@ -51,7 +58,12 @@ public class MeasurementUnitService {
     }
 
     @Transactional
-    public void delete(Long id) {
+    public void delete(Long id) throws SQLIntegrityConstraintViolationException {
+
+        if (skuRepository.existsByMeasurementUnitId(id)) {
+            throw new SQLIntegrityConstraintViolationException("Cannot delete measument unit because it is associated with a product");
+        }
+
         MeasurementUnit measurementUnit = this.read(id);
         measurementUnitRepository.delete(measurementUnit);
     }
